% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Skifti_view.R
\name{save_skeleton}
\alias{save_skeleton}
\title{Create png from mask and data in Nifti format}
\usage{
save_skeleton(
  mask,
  data,
  img_hdr,
  output,
  legend_title,
  scale,
  keep_temp = FALSE,
  palette = "lajolla",
  verbose = FALSE
)
}
\arguments{
\item{mask}{Intensity data in Nifti object format}

\item{data}{Skeleton at same imaging space as the data, in Nifti format}

\item{img_hdr}{Nifti header object}

\item{output}{Output PNG filename}

\item{legend_title}{Title to be shown}

\item{scale}{scaling for intensity values, tune for better color depth}

\item{keep_temp}{TRUE/FALSE(default) to keep temporary png images}

\item{palette}{color palette}

\item{verbose}{TRUE/FALSE(default), for verbosity}
}
\value{
No output, as resuts are saved to a png file
}
\description{
Skeleton mask and corresponding image intensity data must be in Nifti format.
The skeleton mask is used to determine the coordinates of intensity data.
}
