% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operationsSkifti.R
\name{subset}
\alias{subset}
\title{Get subset of Skifti data}
\usage{
subset(Skifti_data, volumes)
}
\arguments{
\item{Skifti_data}{Skifti data object}

\item{volumes}{selection}
}
\value{
Skifti data object of subset
}
\description{
Get subset of Skifti data
}
